Move AndroidLogcatFakeDevice to its own class
