+++ com.unity.mobile.android-logcat/Tests/Editor/AndroidLogcatFakeDevice.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c729e8768ea645418d210874a714ac3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
